Create Preferences module
